requirements.txt
python-dotenv
mira-sdk
numpy
pandas
matplotlib 
streamlit
beautifulsoup4
requests
selenium
youtube-transcript-api